# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Recruit API",
        "description": "採用広報サイト + 応募管理ダッシュボード API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/applications": {
            "post": {
                "tags": [
                    "Applications"
                ],
                "summary": "応募送信（公開）",
                "description": "公開側: 応募送信",
                "operationId": "2b57cdaaa2f01ca15e070687f10ad613",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "応募完了"
                    },
                    "422": {
                        "description": "バリデーションエラー"
                    }
                }
            }
        },
        "/admin/applications": {
            "get": {
                "tags": [
                    "Admin/Applications"
                ],
                "summary": "応募一覧（管理）",
                "description": "管理側: 応募一覧",
                "operationId": "af9a14825fd6f3f1e7fe972ed829cad0",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "keyword",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "応募一覧"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/applications/{application}": {
            "get": {
                "tags": [
                    "Admin/Applications"
                ],
                "summary": "応募詳細（管理）",
                "description": "管理側: 応募詳細",
                "operationId": "e6b6412337c1fb1b9d5bfaa10e7a1b98",
                "parameters": [
                    {
                        "name": "application",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "応募詳細"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/applications/{application}/status": {
            "put": {
                "tags": [
                    "Admin/Applications"
                ],
                "summary": "応募ステータス更新",
                "description": "管理側: ステータス更新",
                "operationId": "5b6f2020eadb537e5ca689b6b5584e59",
                "parameters": [
                    {
                        "name": "application",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "unread",
                                            "reviewing",
                                            "interviewing",
                                            "rejected",
                                            "accepted"
                                        ]
                                    },
                                    "admin_note": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "更新成功"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/applications/export": {
            "get": {
                "tags": [
                    "Admin/Applications"
                ],
                "summary": "応募 CSV エクスポート",
                "description": "管理側: CSV エクスポート",
                "operationId": "2088c1fdb04e60bb98f944b11195b374",
                "responses": {
                    "200": {
                        "description": "CSV ファイル"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "ログイン",
                "operationId": "67d6d0fdd54fba132e1b6f55ca254183",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ログイン成功"
                    },
                    "401": {
                        "description": "認証失敗"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "ログアウト",
                "operationId": "28d48c13fca984400ac8de99f54ee13a",
                "responses": {
                    "200": {
                        "description": "ログアウト成功"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "ログインユーザー情報",
                "operationId": "bc2032a379aa97707ccc61511897c120",
                "responses": {
                    "200": {
                        "description": "ユーザー情報"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}